refactor(agent-hook): simplify fix session lifecycle
Fix ownership is short-lived coordination state, but the first implementation retained terminal records and coupled command routing to prose text. That made state grow and made ordinary hook saves more expensive.

Keep only active owners. Render the completion instruction from typed response data and the running hook binary, and preserve per-entry rollback.
