+++ cmd/roborev/agent_hook_fix_reason.md.gotmpl
@@ -1,0 +1,3 @@
+{{.Reason}}
+
+After completing this Agent Hook fix, run `{{.Command}}`.
